% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census-data.R
\docType{data}
\name{censusData}
\alias{censusData}
\title{Census Demographic Information}
\format{
A data frame with 8850 observations with 5 columns.
\describe{
  \item{index}{numerical index for full dataframe}
  \item{Labels}{the meaning to the value in the estimate column}
  \item{Year}{the year of the given information}
  \item{County}{the county of Arizona the value came from}
  \item{Estimate}{the estimated population value for the given label}
}
}
\source{
\url{https://data.census.gov/cedsci/table}
}
\usage{
censusData
}
\description{
This data has information on the top ten populated counties in Arizona. There
are estimates for different demographic information on the population of
Arizona as collected from 2010 to 2019 in American Community Surveys data
profiles.
}
\keyword{datasets}
